LayoutTests/http/tests/security/canvas-remote-read-redirect-to-remote-image-expected.txt

CONSOLE MESSAGE: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
This tests that drawing an image that redirects to a remote image to a canvas taints the canvas

PASS: getImageData failed. Canvas tainted.
 
